--- src/main/java/com/example/challengewithmebe/user/domain/Member.java
@@ -1,23 +1,0 @@
-package com.example.challengewithmebe.user.domain;
-
-import com.example.challengewithmebe.global.domain.BaseEntity;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Member extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    private String imgUrl; //프로필 이미지 링크
-    private String email;
-}
